Fully implemented rigidbody as a component. Improved component interface.
