+++ src/core/game_object/game_object.cpp
@@ -12,7 +12,9 @@
 
 #include "glm/gtc/quaternion.hpp"
 #include "components/component.h"
+#include "components/rigid_body_component.h"
 #include "src/core/engine.h"
+#include "src/core/time.h"
 #include "src/core/identifier/identifier_manager.h"
 #include "src/physics/physics.h"
 #include "src/physics/physics_filters.h"
@@ -52,13 +54,6 @@ GameObject::~GameObject()
         pRenderReference->releaseInstanceIndex(this);
         pRenderReference = nullptr;
     }
-
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        if (physics::Physics* physics = physics::Physics::Get()) {
-            physics->removeRigidBody(this);
-            bodyId = JPH::BodyID(JPH::BodyID::cMaxBodyIndex);
-        }
-    }
 }
 
 void GameObject::destroy()
@@ -90,7 +85,7 @@ void GameObject::setName(std::string newName)
 void GameObject::beginPlay()
 {
     for (const auto& component : components) {
-        component->beginPlay(this);
+        component->beginPlay();
     }
     bHasBegunPlay = true;
 }
@@ -220,8 +215,8 @@ void GameObject::setLocalPosition(const glm::vec3 localPosition)
 {
     transform.setPosition(localPosition);
     dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
     }
 }
 
@@ -229,8 +224,8 @@ void GameObject::setLocalRotation(const glm::quat localRotation)
 {
     transform.setRotation(localRotation);
     dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
     }
 }
 
@@ -238,8 +233,8 @@ void GameObject::setLocalScale(const glm::vec3 localScale)
 {
     transform.setScale(localScale);
     dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
     }
 }
 
@@ -252,8 +247,8 @@ void GameObject::setLocalTransform(const Transform& newLocalTransform)
 {
     transform = newLocalTransform;
     dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
     }
 }
 
@@ -320,35 +315,8 @@ void GameObject::setGlobalTransform(const Transform& newGlobalTransform)
     }
 
     dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
-    }
-}
-
-void GameObject::translate(const glm::vec3 translation)
-{
-    transform.translate(translation);
-    dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
-    }
-}
-
-void GameObject::rotate(const glm::quat rotation)
-{
-    transform.rotate(rotation);
-    dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
-    }
-}
-
-void GameObject::rotateAxis(const float angle, const glm::vec3& axis)
-{
-    transform.rotateAxis(angle, axis);
-    dirty();
-    if (bodyId.GetIndex() != JPH::BodyID::cMaxBodyIndex) {
-        physics::Physics::Get()->setPositionAndRotation(bodyId, getGlobalPosition(), getGlobalRotation());
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
     }
 }
 
@@ -367,6 +335,46 @@ void GameObject::setGlobalTransformFromPhysics(const glm::vec3& position, const 
     dirty();
 }
 
+void GameObject::translate(const glm::vec3 translation)
+{
+    transform.translate(translation);
+    dirty();
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
+    }
+}
+
+void GameObject::rotate(const glm::quat rotation)
+{
+    transform.rotate(rotation);
+    dirty();
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
+    }
+}
+
+void GameObject::rotateAxis(const float angle, const glm::vec3& axis)
+{
+    transform.rotateAxis(angle, axis);
+    dirty();
+    if (rigidbodyComponent) {
+        rigidbodyComponent->setPhysicsTransformFromGame(getGlobalPosition(), getGlobalRotation());
+    }
+}
+
+
+bool GameObject::canAddComponent(const std::string_view componentType)
+{
+    for (const auto& _component : components) {
+        if (componentType == _component->getComponentType()) {
+            fmt::print("Attempted to add a component of the same type to a gameobject. This is not supported at this time.\n");
+            return false;
+        }
+    }
+
+    return true;
+}
+
 void GameObject::addComponent(std::unique_ptr<components::Component> component)
 {
     if (!component) { return; }
@@ -377,10 +385,14 @@ void GameObject::addComponent(std::unique_ptr<components::Component> component)
             return;
         }
     }
+    component->setOwner(this);
     components.push_back(std::move(component));
 
+
+    cacheRigidbody();
+
     if (bHasBegunPlay) {
-        components.back()->beginPlay(this);
+        components.back()->beginPlay();
     }
 }
 
@@ -392,12 +404,27 @@ void GameObject::destroyComponent(components::Component* component)
         return comp.get() == component;
     });
 
+    if (component == rigidbodyComponent) {
+        rigidbodyComponent = nullptr;
+        cacheRigidbody();
+    }
+
     if (it != components.end()) {
         component->beginDestroy();
         components.erase(it);
     }
     else {
         fmt::print("Attempted to remove a component that does not belong to this gameobject.\n");
+    }
+}
+
+void GameObject::cacheRigidbody()
+{
+    for (const auto& comp : components) {
+        if (const auto rb = dynamic_cast<components::RigidBodyComponent*>(comp.get())) {
+            rigidbodyComponent = rb;
+            break;
+        }
     }
 }
 
@@ -441,270 +468,22 @@ void GameObject::selectedRenderImgui()
                     }
                 }
 
-                // IPhysicsBody
-                if (ImGui::CollapsingHeader("Physics", ImGuiTreeNodeFlags_DefaultOpen)) {
-                    const JPH::BodyID bodyId = getPhysicsBodyId();
-
-                    if (bodyId.GetIndex() == JPH::BodyID::cMaxBodyIndex) {
-                        static int selectedShape = 0;
-                        const char* shapes[] = {"Box", "Sphere", "Capsule", "Cylinder"};
-                        ImGui::Combo("Shape", &selectedShape, shapes, 4);
-                        static bool useUnitShape = true;
-                        glm::vec3 shapeParams = glm::vec3(1.0f);
-
-                        // Shape Generation
-                        {
-                            ImGui::Checkbox("Use Unit Shape", &useUnitShape);
-                            if (!useUnitShape) {
-                                switch (selectedShape) {
-                                    case 0: // Box
-                                    {
-                                        static glm::vec3 boxHalfExtents(1.0f);
-                                        ImGui::Text("Half Extents");
-                                        ImGui::DragFloat3("##BoxHalfExtents", &boxHalfExtents.x, 0.1f, 0.1f, 100.0f);
-                                        shapeParams = boxHalfExtents;
-                                        break;
-                                    }
-                                    case 1: // Sphere
-                                    {
-                                        static float radius = 1.0f;
-                                        ImGui::Text("Radius");
-                                        ImGui::DragFloat("##SphereRadius", &radius, 0.1f, 0.1f, 100.0f);
-                                        shapeParams = glm::vec3(radius, 0.0f, 0.0f);
-                                        break;
-                                    }
-                                    case 2: // Capsule
-                                    {
-                                        static float radius = 1.0f;
-                                        static float halfHeight = 1.0f;
-                                        ImGui::Text("Radius");
-                                        ImGui::DragFloat("##CapsuleRadius", &radius, 0.1f, 0.1f, 100.0f);
-                                        ImGui::Text("Half Height");
-                                        ImGui::DragFloat("##CapsuleHeight", &halfHeight, 0.1f, 0.1f, 100.0f);
-                                        shapeParams = glm::vec3(radius, halfHeight, 0.0f);
-                                        break;
-                                    }
-                                    case 3: // Cylinder
-                                    {
-                                        static float radius = 1.0f;
-                                        static float halfHeight = 1.0f;
-                                        ImGui::Text("Radius");
-                                        ImGui::DragFloat("##CylinderRadius", &radius, 0.1f, 0.1f, 100.0f);
-                                        ImGui::Text("Half Height");
-                                        ImGui::DragFloat("##CylinderHeight", &halfHeight, 0.1f, 0.1f, 100.0f);
-                                        shapeParams = glm::vec3(radius, halfHeight, 0.0f);
-                                        break;
-                                    }
-                                    default:
-                                        shapeParams = glm::vec3(1.0f);
-                                        break;
-                                }
-                            }
-                            else {
-                                switch (selectedShape) {
-                                    case 0:
-                                        shapeParams = physics::UNIT_CUBE;
-                                        break;
-                                    case 1:
-                                        shapeParams = physics::UNIT_SPHERE;
-                                        break;
-                                    case 2:
-                                        shapeParams = physics::UNIT_CAPSULE;
-                                        break;
-                                    case 3:
-                                        shapeParams = physics::UNIT_CYLINDER;
-                                        break;
-                                    default:
-                                        shapeParams = glm::vec3(1.0f);
-                                        break;
-                                }
-                            }
-                        }
-
-                        static auto motionType = JPH::EMotionType::Dynamic;
-                        const char* motionTypes[] = {"Static", "Kinematic", "Dynamic"};
-                        int currentType = static_cast<int>(motionType);
-                        ImGui::Combo("Motion Type", &currentType, motionTypes, 3);
-                        motionType = static_cast<JPH::EMotionType>(currentType);
-
-                        const char* layers[] = {"Non-Moving", "Moving", "Player", "Terrain"};
-                        static JPH::ObjectLayer layer = physics::Layers::MOVING;
-                        int currentLayer = layer;
-                        ImGui::Combo("Layer", &currentLayer, layers, 4);
-                        layer = static_cast<JPH::ObjectLayer>(currentLayer);
-
-
-                        if (ImGui::Button("Add Rigidbody")) {
-                            JPH::EShapeSubType shapeType;
-
-                            switch (selectedShape) {
-                                case 0:
-                                    shapeType = JPH::EShapeSubType::Box;
-                                    break;
-                                case 1:
-                                    shapeType = JPH::EShapeSubType::Sphere;
-                                    break;
-                                case 2:
-                                    shapeType = JPH::EShapeSubType::Capsule;
-                                    break;
-                                case 3:
-                                    shapeType = JPH::EShapeSubType::Cylinder;
-                                    break;
-                                default: shapeType = JPH::EShapeSubType::Box;
-                                    break;
-                            }
-
-                            physics::Physics::Get()->setupRigidbody(
-                                this,
-                                shapeType,
-                                shapeParams,
-                                motionType,
-                                layer
-                            );
-                        }
-                    }
-                    else {
-                        ImGui::Text("Body Id: %u", bodyId);
-
-                        ImGui::Separator();
-
-                        // Layer
-                        {
-                            JPH::ObjectLayer layer = physics::PhysicsUtils::getObjectLayer(bodyId);
-                            if (layer < physics::Layers::NUM_LAYERS) {
-                                auto layerName = physics::Layers::layerNames[layer];
-                                ImGui::Text("Layer: %s", layerName);
-                            }
-                            else {
-                                ImGui::Text("Layer: Invalid Layer Found greater than NUM_LAYERS");
-                            }
-                        }
-
-                        ImGui::Separator();
-
-                        // Shape
-                        {
-                            if (physics::PhysicsObject* physicsObject = physics::Physics::Get()->getPhysicsObject(bodyId)) {
-                                const JPH::ShapeRefC shape = physicsObject->shape;
-                                ImGui::Text("Shape: ");
-                                ImGui::SameLine();
-
-                                switch (physicsObject->shape->GetSubType()) {
-                                    case JPH::EShapeSubType::Box:
-                                    {
-                                        const auto box = static_cast<const JPH::BoxShape*>(shape.GetPtr());
-                                        ImGui::TextColored(ImVec4(0.4f, 0.8f, 0.4f, 1.0f), "Box");
-                                        const JPH::Vec3 halfExtent = box->GetHalfExtent();
-                                        ImGui::Columns(2);
-                                        ImGui::Text("Half Extents");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("X: %.2f  Y: %.2f  Z: %.2f", halfExtent.GetX(), halfExtent.GetY(), halfExtent.GetZ());
-                                        ImGui::Columns(1);
-                                        break;
-                                    }
-                                    case JPH::EShapeSubType::Sphere:
-                                    {
-                                        const auto sphere = static_cast<const JPH::SphereShape*>(shape.GetPtr());
-                                        ImGui::TextColored(ImVec4(0.4f, 0.8f, 0.4f, 1.0f), "Sphere");
-                                        ImGui::Columns(2);
-                                        ImGui::Text("Radius");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("%.2f", sphere->GetRadius());
-                                        ImGui::Columns(1);
-                                        break;
-                                    }
-                                    case JPH::EShapeSubType::Capsule:
-                                    {
-                                        const auto capsule = static_cast<const JPH::CapsuleShape*>(shape.GetPtr());
-                                        ImGui::TextColored(ImVec4(0.4f, 0.8f, 0.4f, 1.0f), "Capsule");
-                                        ImGui::Columns(2);
-                                        ImGui::Text("Radius");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("%.2f", capsule->GetRadius());
-                                        ImGui::NextColumn();
-                                        ImGui::Text("Half Height");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("%.2f", capsule->GetHalfHeightOfCylinder());
-                                        ImGui::Columns(1);
-                                        break;
-                                    }
-                                    case JPH::EShapeSubType::Cylinder:
-                                    {
-                                        const auto cylinder = static_cast<const JPH::CylinderShape*>(shape.GetPtr());
-                                        ImGui::TextColored(ImVec4(0.4f, 0.8f, 0.4f, 1.0f), "Cylinder");
-                                        ImGui::Columns(2);
-                                        ImGui::Text("Radius");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("%.2f", cylinder->GetRadius());
-                                        ImGui::NextColumn();
-                                        ImGui::Text("Half Height");
-                                        ImGui::NextColumn();
-                                        ImGui::Text("%.2f", cylinder->GetHalfHeight());
-                                        ImGui::Columns(1);
-                                        break;
-                                    }
-                                    default:
-                                        ImGui::TextColored(ImVec4(0.8f, 0.4f, 0.4f, 1.0f), "Unknown");
-                                        break;
-                                }
-                            }
-                        }
-
-                        ImGui::Separator();
-
-                        // MotionType
-                        {
-                            JPH::EMotionType currentMotionType = physics::Physics::Get()->getMotionType(this);
-                            if (currentMotionType == JPH::EMotionType::Static) {
-                                ImGui::Text("Motion Type: Static (Unable to change)");
-                            }
-                            else {
-                                const char* motionTypes[] = {"Kinematic", "Dynamic"};
-                                int currentType = static_cast<int>(currentMotionType) - 1;
-
-                                if (ImGui::Combo("Motion Type", &currentType, motionTypes, 2)) {
-                                    const JPH::EMotionType newMotionType = static_cast<JPH::EMotionType>(currentType + 1);
-                                    switch (newMotionType) {
-                                        case JPH::EMotionType::Kinematic:
-                                            physics::Physics::Get()->setMotionType(
-                                                this,
-                                                static_cast<JPH::EMotionType>(currentType + 1),
-                                                JPH::EActivation::DontActivate
-                                            );
-                                            break;
-                                        case JPH::EMotionType::Dynamic:
-                                            physics::Physics::Get()->setMotionType(
-                                                this,
-                                                static_cast<JPH::EMotionType>(currentType + 1),
-                                                JPH::EActivation::Activate
-                                            );
-                                            break;
-                                        default:
-                                            break;
-                                    }
-                                    physics::PhysicsUtils::resetVelocity(bodyId);
-                                }
-
-                                glm::vec3 velocity = physics::PhysicsUtils::getLinearVelocity(bodyId);
-                                ImGui::Text("Velocity: %.2f, %.2f, %.2f", velocity.x, velocity.y, velocity.z);
-
-                                if (ImGui::Button("Reset Velocity")) {
-                                    physics::PhysicsUtils::resetVelocity(bodyId);
-                                }
-
-
-                                if (ImGui::Button("Remove Rigidbody")) {
-                                    physics::Physics::Get()->releaseRigidbody(this);
-                                }
-                            }
-                        }
-                    }
-                }
 
                 ImGui::EndTabItem();
             }
 
             if (ImGui::BeginTabItem("Components")) {
+                static std::string failedComponentMessage;
+                static float failedComponentTimeLeft{0};
+
+
+                if (failedComponentTimeLeft > 0) {
+                    if (!failedComponentMessage.empty()) {
+                        ImGui::TextColored(ImVec4(1.0f, 0.0f, 0.0f, 1.0f), "%s", failedComponentMessage.c_str());
+                    }
+                    failedComponentTimeLeft -= Time::Get().getDeltaTime();
+                }
+
                 if (ImGui::Button("Add Component")) {
                     ImGui::OpenPopup("AddComponentPopup");
                 }
@@ -716,10 +495,17 @@ void GameObject::selectedRenderImgui()
                     const auto& creators = components::ComponentFactory::getInstance().getComponentCreators();
                     for (const auto& type : creators | std::views::keys) {
                         if (ImGui::Selectable(type.data())) {
-                            auto newComponent = components::ComponentFactory::getInstance().createComponent(type, "New " + std::string(type));
-                            if (newComponent) {
-                                addComponent(std::move(newComponent));
+                            if (!canAddComponent(type)) {
+                                failedComponentMessage = "Cannot add another " + std::string(type) + " component";
+                                failedComponentTimeLeft = 7.5f;
                             }
+                            else {
+                                auto newComponent = components::ComponentFactory::getInstance().createComponent(type, "New " + std::string(type));
+                                if (newComponent) {
+                                    addComponent(std::move(newComponent));
+                                }
+                            }
+
                             ImGui::CloseCurrentPopup();
                         }
                     }
@@ -747,13 +533,21 @@ void GameObject::selectedRenderImgui()
                     }
 
                     ImGui::SameLine();
+                    const char* componentType = component->getComponentType().data();
                     if (ImGui::Button(componentName.c_str(), ImVec2(-1, 0))) {
                         component->openRenderImgui();
-                        ImGui::OpenPopup("Component Editor");
+                        ImGui::OpenPopup(componentType);
                     }
 
-                    if (ImGui::BeginPopupModal("Component Editor", nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-                        ImGui::Text("Editing %s Component", component->getComponentType().data());
+
+                    ImGui::PushStyleColor(ImGuiCol_PopupBg, ImVec4(0.75f, 0.75f, 0.75f, 1.0f)); // Dark gray
+                    if (ImGui::BeginPopupModal(componentType, nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
+                        static char objectName[256] = "";
+                        if (ImGui::InputText("Component Name", objectName, sizeof(objectName))) {
+                            // ReSharper disable once CppDFAUnusedValue
+                            componentName = objectName;
+                        }
+                        component->setComponentName(objectName);
                         ImGui::Separator();
 
                         component->updateRenderImgui();
@@ -778,6 +572,7 @@ void GameObject::selectedRenderImgui()
 
                         ImGui::EndPopup();
                     }
+                    ImGui::PopStyleColor();
 
                     ImGui::PopID();
                 }
